Scenarios/5. Lost City of Z/re_archaeology_mvp/requirements.txt
fastapi==0.104.1
uvicorn==0.23.2
psycopg2-binary==2.9.9
geopandas==0.14.0
numpy==1.26.0
scipy==1.11.3
pandas==2.1.1
earthengine-api==0.1.370
google-cloud-storage==2.10.0
google-auth==2.23.0
pydantic==2.4.2
pydantic-settings==2.0.3
python-dotenv==1.0.0
pytest==7.4.2
httpx==0.25.0
shapely==2.0.2
rasterio==1.3.8
# gdal is installed separately in Dockerfile to match system version
python-multipart==0.0.6  # Required for form data handling
redis==5.0.1
celery==5.3.4
SQLAlchemy==2.0.22
SQLAlchemy-Utils==0.41.1
GeoAlchemy2==0.14.1
